refactor: move consistency checking to typstyle-consistency crate and merge test entries
