+++ crates/typstyle-consistency/Cargo.toml
@@ -1,0 +1,27 @@
+[package]
+name = "typstyle-consistency"
+version.workspace = true
+rust-version.workspace = true
+authors.workspace = true
+edition.workspace = true
+publish = false
+
+[lib]
+test = false
+doctest = false
+bench = false
+doc = false
+
+[dependencies]
+anyhow.workspace = true
+ecow.workspace = true
+itertools.workspace = true
+walkdir.workspace = true
+
+pretty_assertions.workspace = true
+reflexo-world.workspace = true
+reflexo-typst.workspace = true
+typst-assets.workspace = true
+typst-render.workspace = true
+typst-pdf.workspace = true
+typst.workspace = true
