+++ crates/typstyle-consistency/src/lib.rs
@@ -1,0 +1,8 @@
+pub mod cmp;
+pub mod compile;
+
+use ecow::EcoVec;
+use std::sync::Arc;
+use typst::{diag::SourceDiagnostic, model::Document};
+
+pub type CompilationResult = Result<Arc<Document>, EcoVec<SourceDiagnostic>>;
