+++ tests/src/assets.rs
@@ -1,29 +1,34 @@
-use ecow::EcoVec;
-use libtest_mimic::{Arguments, Failed, Trial};
-use reflexo_typst::CompileDriver;
-use reflexo_world::{config::CompileOpts, EntryOpts, ShadowApi, TypstSystemUniverse};
-use std::{
-    borrow::Cow,
-    env,
-    error::Error,
-    ffi::OsStr,
-    fs,
-    marker::PhantomData,
-    path::{Path, PathBuf},
-    sync::Arc,
-};
-use typst::{diag::SourceDiagnostic, layout::Page, model::Document};
+use libtest_mimic::{Failed, Trial};
+use std::{env, error::Error, ffi::OsStr, fs, path::Path};
 use typstyle_core::Typstyle;
-
-fn main() -> Result<(), Box<dyn Error>> {
-    let args = Arguments::from_args();
-    let tests = collect_tests()?;
-    libtest_mimic::run(&args, tests).exit();
-}
 
 /// Creates one test for each `.typ` file in the current directory or
 /// sub-directories of the current directory.
-fn collect_tests() -> Result<Vec<Trial>, Box<dyn Error>> {
+pub fn collect_tests() -> Result<Vec<Trial>, Box<dyn Error>> {
+    fn make_test(path: &Path, name: &str, width: usize) -> Trial {
+        let path = path.to_path_buf();
+        Trial::test(format!("{name} - {width}char"), move || {
+            check_file(&path, width)
+        })
+        .with_kind("typst")
+    }
+
+    fn make_convergence_test(path: &Path, name: &str, width: usize) -> Trial {
+        let path = path.to_path_buf();
+        Trial::test(format!("{name} - convergence - {width}char"), move || {
+            check_convergence(&path, width)
+        })
+    }
+
+    #[cfg(feature = "consistency")]
+    fn make_consistency_test(path: &Path, name: &str, width: usize) -> Trial {
+        let path = path.to_path_buf();
+        Trial::test(
+            format!("{name} - output consistency - {width}char"),
+            move || check_output_consistency(&path, width),
+        )
+    }
+
     fn visit_dir(path: &Path, tests: &mut Vec<Trial>) -> Result<(), Box<dyn Error>> {
         for entry in fs::read_dir(path)? {
             let entry = entry?;
@@ -37,78 +42,21 @@ fn collect_tests() -> Result<Vec<Trial>, Box<dyn Error>> {
                         .strip_prefix(env::current_dir()?)?
                         .display()
                         .to_string();
-                    let test_0 = {
-                        let path = path.clone();
-                        Trial::test(format!("{} - 0char", name), move || check_file(&path, 0))
-                            .with_kind("typst")
-                    };
-                    let test_40 = {
-                        let path: std::path::PathBuf = path.clone();
-                        Trial::test(format!("{} - 40char", name), move || check_file(&path, 40))
-                            .with_kind("typst")
-                    };
-                    let test_80 = {
-                        let path = path.clone();
-                        Trial::test(format!("{} - 80char", name), move || check_file(&path, 80))
-                            .with_kind("typst")
-                    };
-                    let test_120 = {
-                        let path = path.clone();
-                        Trial::test(format!("{} - 120char", name), move || {
-                            check_file(&path, 120)
-                        })
-                        .with_kind("typst")
-                    };
-                    let test_convergence_0 = {
-                        let path = path.clone();
-                        Trial::test(format!("{} - convergence - 0char", name), move || {
-                            check_convergence(&path, 0)
-                        })
-                    };
-                    let test_convergence_40 = {
-                        let path = path.clone();
-                        Trial::test(format!("{} - convergence - 40char", name), move || {
-                            check_convergence(&path, 40)
-                        })
-                    };
-                    let test_convergence_80 = {
-                        let path = path.clone();
-                        Trial::test(format!("{} - convergence - 80char", name), move || {
-                            check_convergence(&path, 80)
-                        })
-                    };
-                    let test_output_consistency_80 = {
-                        let path = path.clone();
-                        Trial::test(
-                            format!("{} - output consistency - 80char", name),
-                            move || check_output_consistency(&path, 80),
-                        )
-                    };
-                    let test_output_consistency_40 = {
-                        let path = path.clone();
-                        Trial::test(
-                            format!("{} - output consistency - 40char", name),
-                            move || check_output_consistency(&path, 40),
-                        )
-                    };
-                    let test_output_consistency_0 = {
-                        let path = path.clone();
-                        Trial::test(
-                            format!("{} - output consistency - 0char", name),
-                            move || check_output_consistency(&path, 0),
-                        )
-                    };
+
                     tests.extend([
-                        test_0,
-                        test_40,
-                        test_80,
-                        test_120,
-                        test_convergence_0,
-                        test_convergence_40,
-                        test_convergence_80,
-                        test_output_consistency_0,
-                        test_output_consistency_40,
-                        test_output_consistency_80,
+                        make_test(&path, &name, 0),
+                        make_test(&path, &name, 40),
+                        make_test(&path, &name, 80),
+                        make_test(&path, &name, 120),
+                        make_convergence_test(&path, &name, 0),
+                        make_convergence_test(&path, &name, 40),
+                        make_convergence_test(&path, &name, 80),
+                    ]);
+                    #[cfg(feature = "consistency")]
+                    tests.extend([
+                        make_consistency_test(&path, &name, 0),
+                        make_consistency_test(&path, &name, 40),
+                        make_consistency_test(&path, &name, 80),
                     ]);
                 }
             } else if file_type.is_dir() {
@@ -155,8 +103,9 @@ fn check_file(path: &Path, width: usize) -> Result<(), Failed> {
         snapshot_path => env::current_dir().unwrap().join("snapshots"),
         snapshot_suffix => format!("{}-{width}", replaced_path),
         input_file => path,
+        prepend_module_to_snapshot => false,
     }, {
-        insta::assert_snapshot!(doc_string);
+        insta::assert_snapshot!("assets__check_file", doc_string);
     });
     Ok(())
 }
@@ -178,122 +127,23 @@ fn check_convergence(path: &Path, width: usize) -> Result<(), Failed> {
     Ok(())
 }
 
-fn compile_typst_src(content: &str) -> Result<Arc<Document>, EcoVec<SourceDiagnostic>> {
-    let root = if cfg!(windows) {
-        PathBuf::from("C:\\")
-    } else {
-        PathBuf::from("/")
-    };
-    let mut univ = TypstSystemUniverse::new(CompileOpts {
-        entry: EntryOpts::new_rooted(root.clone(), Some(PathBuf::from("/main.typ"))),
-        with_embedded_fonts: typst_assets::fonts().map(Cow::Borrowed).collect(),
-        ..Default::default()
-    })
-    .unwrap()
-    .with_entry_file(root.join("main.typ"));
-    univ.map_shadow(&root.join("main.typ"), content.as_bytes().into())
-        .unwrap();
-    let compiler = PhantomData;
-    let mut driver = CompileDriver::new(compiler, univ);
-    driver.compile(&mut Default::default()).map(|x| x.output)
-}
-
+#[cfg(feature = "consistency")]
 fn check_output_consistency(path: &Path, width: usize) -> Result<(), Failed> {
-    let content = fs::read(path).map_err(|e| format!("Cannot read file: {e}"))?;
+    use typstyle_consistency::{
+        cmp::compare_docs,
+        compile::{compile_universe, make_universe},
+    };
 
+    let content = fs::read(path).map_err(|e| format!("Cannot read file: {e}"))?;
     // Check that the file is valid UTF-8
     let content = String::from_utf8(content)
         .map_err(|_| "The file's contents are not a valid UTF-8 string!")?;
     let content = remove_crlf(content);
     let formatted_src = Typstyle::new_with_content(content.clone(), width).pretty_print();
-    let doc = compile_typst_src(&content);
-    let formatted_doc = compile_typst_src(&formatted_src);
+
+    let doc = compile_universe(make_universe(&content)?).0;
+    let formatted_doc = compile_universe(make_universe(&formatted_src)?).0;
     compare_docs(doc, formatted_doc)?;
-    Ok(())
-}
 
-fn compare_docs(
-    doc: Result<Arc<Document>, EcoVec<SourceDiagnostic>>,
-    formatted_doc: Result<Arc<Document>, EcoVec<SourceDiagnostic>>,
-) -> Result<(), Failed> {
-    match (doc, formatted_doc) {
-        (Ok(doc), Ok(formatted_doc)) => {
-            pretty_assertions::assert_eq!(
-                doc.pages.len(),
-                formatted_doc.pages.len(),
-                "The page counts are not consistent"
-            );
-            pretty_assertions::assert_eq!(
-                doc.info.title,
-                formatted_doc.info.title,
-                "The titles are not consistent"
-            );
-            pretty_assertions::assert_eq!(
-                doc.info.author,
-                formatted_doc.info.author,
-                "The authors are not consistent"
-            );
-            pretty_assertions::assert_eq!(
-                doc.info.keywords,
-                formatted_doc.info.keywords,
-                "The keywords are not consistent"
-            );
-
-            for (i, (doc, formatted_doc)) in
-                doc.pages.iter().zip(formatted_doc.pages.iter()).enumerate()
-            {
-                let png = typst_render::render(
-                    &Page {
-                        frame: doc.frame.clone(),
-                        fill: typst::foundations::Smart::Auto,
-                        numbering: None,
-                        number: i,
-                    },
-                    2.0,
-                );
-                let formatted_png = typst_render::render(
-                    &Page {
-                        frame: formatted_doc.frame.clone(),
-                        fill: typst::foundations::Smart::Auto,
-                        numbering: None,
-                        number: i,
-                    },
-                    2.0,
-                );
-                if png != formatted_png {
-                    // save both to tmp path and report error
-                    let tmp_dir = env::temp_dir();
-                    let png_path = tmp_dir.join(format!("{}-{}.png", i, "original"));
-                    let formatted_png_path = tmp_dir.join(format!("{}-{}.png", i, "formatted"));
-                    png.save_png(&png_path).unwrap();
-                    formatted_png.save_png(&formatted_png_path).unwrap();
-                    return Err(Failed::from(format!(
-                        "The output are not consistent for page {}, original png path: {}, formatted png path: {}",
-                        i, png_path.display(), formatted_png_path.display()
-                    )));
-                }
-            }
-        }
-        (Err(e1), Err(e2)) => {
-            pretty_assertions::assert_eq!(
-                e1.len(),
-                e2.len(),
-                "The error counts are not consistent"
-            );
-            for (e1, e2) in e1.iter().zip(e2.iter()) {
-                pretty_assertions::assert_eq!(
-                    e1.message,
-                    e2.message,
-                    "The error messages are not consistent after formatting"
-                );
-            }
-        }
-        (res1, res2) => {
-            return Err(Failed::from(format!(
-                "One of the documents failed to compile: {:#?} {:#?}",
-                res1, res2
-            )));
-        }
-    }
     Ok(())
 }
