+++ tests/src/tests.rs
@@ -1,0 +1,16 @@
+mod assets;
+#[cfg(feature = "consistency")]
+#[path = "repo-e2e.rs"]
+mod repo_e2e;
+
+use libtest_mimic::Arguments;
+use std::error::Error;
+
+fn main() -> Result<(), Box<dyn Error>> {
+    #[allow(unused_mut)]
+    let mut tests = assets::collect_tests()?;
+    #[cfg(feature = "consistency")]
+    tests.append(&mut repo_e2e::collect_tests());
+
+    libtest_mimic::run(&Arguments::from_args(), tests).exit();
+}
